Comments can be created and display in (almost) real-time
In a single instance of the app, you can create a comment on an article and it displays in the article spotlight in real-time. However, if you have two parallel instances, you have to close the spotlight in the second instance and re-open the spotlight to see the new comment.